Fix #1
When Anyconnect prompt is opened the "Connect" button is already
selected.
This eliminates the need for using Sikuli to click the button.

Import keyboard.press() to hit carriage return instead.
This also eliminates the need to bring the prompt back to top,
because the press function occurs 1 sec after proc opens.
Removed Sikuli functionality from script.